capdev: Separate *nix specific code to prepare Windows support
